requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.5
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
discord-py==2.3.2
    # via -r requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via
    #   requests
    #   yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
requests==2.31.0
    # via -r requirements.in
urllib3==2.0.4
    # via requests
yarl==1.9.2
    # via aiohttp
